Bug for degrees of bimodal network instead of using numpy.str convert them to int

+++ runwesim.py
@@ -44,8 +44,8 @@ def job_to_cluster(foldername,parameters,Istar):
     if prog == 'bd':
         np.save('parameters.npy',parameters)
         # G = nx.complete_graph(N)
-        d1_in, d1_out, d2_in, d2_out = int(k * (1 - eps_din)), int(k * (1 - eps_dout)), int(k * (1 + eps_din)), int(
-            k * (1 + eps_dout))
+        d1_in, d1_out, d2_in, d2_out = int(int(k) * (1 - float(eps_din))), int(int(k) * (1 - float(eps_dout))), int(int(k) * (1 + float(eps_din))), int(
+            int(k) * (1 + float(eps_dout)))
         Beta = Beta_avg / (1 + eps_din * eps_dout)  # This is so networks with different std will have the reproduction number
     dir_path = os.path.dirname(os.path.realpath(__file__))
     for i in range(int(number_of_networks)):
